Parse response settings as dictionary
There's a bunch of setting messages in the response, each of which
containing a name and a value. This can be easily parsed as a
dictionary.

Also literally taken from LineageOS's nibbler and adapted because this
thing isn't a Discord bot.

While I'm here, get the update title. That way, it's easier to
understand what it is we just got.